resolves #16 - refactored EventIcon and EventIconRow into view.calendar.dart: semantically makes more sense
